Drop redundant extsh in RollingBarrel_update rotZ stores
RollingBarrel_update 97.3->97.9: storing a float->int conversion into
the s16 anim.rotZ via (s16) instead of (s32) drops the spurious extsh
the (s32) cast inserted before the sth (recipe #41/#83c). Both the
rising and falling pitch branches.

+++ src/main/dll/dll_017D_rollingbarrel.c
@@ -219,7 +219,7 @@ void RollingBarrel_update(int obj)
             if (state->pitchRising != 0)
             {
                 ((GameObject*)obj)->anim.rotZ =
-                    (s32)(lbl_803E44A8 * timeDelta + (f32)(int)((GameObject*)obj)->anim.rotZ);
+                    (s16)(lbl_803E44A8 * timeDelta + (f32)(int)((GameObject*)obj)->anim.rotZ);
                 if (((GameObject*)obj)->anim.rotZ > 0x5000)
                 {
                     state->pitchRising = 0;
@@ -228,7 +228,7 @@ void RollingBarrel_update(int obj)
             else
             {
                 ((GameObject*)obj)->anim.rotZ =
-                    (s32) - (lbl_803E44A8 * timeDelta - (f32)(int)((GameObject*)obj)->anim.rotZ);
+                    (s16) - (lbl_803E44A8 * timeDelta - (f32)(int)((GameObject*)obj)->anim.rotZ);
                 if (((GameObject*)obj)->anim.rotZ < 0x3a00)
                 {
                     state->pitchRising = 1;
